feat: move csv generation to server side
